reactor shenyu agent logging #3064 (#3073)
* reactor shenyu agent logging #3064

* reactor shenyu agent logging #3064
