+++ shenyu-agent/shenyu-agent-plugins/pom.xml
@@ -29,7 +29,6 @@
     <modules>
         <module>shenyu-agent-plugin-tracing</module>
         <module>shenyu-agent-plugin-metrics</module>
-        <module>shenyu-agent-plugin-logging</module>
     </modules>
 
     <build>
