--- shenyu-agent/shenyu-agent-plugins/shenyu-agent-plugin-logging/shenyu-agent-plugin-logging-common/src/main/java/org/apache/shenyu/agent/plugin/logging/common/DefaultLogCollector.java
@@ -1,45 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shenyu.agent.plugin.logging.common;
-
-import org.apache.shenyu.agent.plugin.logging.LogCollector;
-import org.apache.shenyu.agent.plugin.logging.LogConsumeClient;
-
-import java.util.Objects;
-
-/**
- * default log collector，depend a LogConsumeClient for consume logs.
- */
-public class DefaultLogCollector extends AbstractLogCollector {
-
-    private static LogCollector instance;
-
-    public DefaultLogCollector(final LogConsumeClient logConsumeClient) {
-        super(logConsumeClient);
-        instance = this;
-    }
-
-    /**
-     * get LogCollector instance.
-     *
-     * @return LogCollector instance
-     */
-    public static LogCollector getInstance() {
-        return Objects.requireNonNull(instance);
-    }
-}
