+++ shenyu-dist/shenyu-agent-dist/src/main/assembly/binary.xml
@@ -55,14 +55,6 @@
             </includes>
         </fileSet>
 
-        <fileSet>
-            <directory>${project.basedir}/../../shenyu-agent/shenyu-agent-plugins/shenyu-agent-plugin-logging/target/plugins</directory>
-            <outputDirectory>./plugins</outputDirectory>
-            <includes>
-                <include>**.jar</include>
-            </includes>
-        </fileSet>
-        
         
     </fileSets>
