--- shenyu-dist/shenyu-agent-dist/src/main/resources/conf/logging-meta.yaml
@@ -1,31 +1,0 @@
-#
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.
-# The ASF licenses this file to You under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
-# the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-logFieldSwitchConfig:
-  requestHeader: true
-  responseHeader: true
-  requestBody: true
-  responseBody: true
-
-globalLogConfig:
-  sampleRate: "1"
-  bufferQueueSize: 50000
-
-logApiSwitchConfigMap:
-  "[/http/test/**]":
-    sampleRate: "0.5"
-    topic: "shenyu-agent-logging-test"
