+++ shenyu-spring-boot-starter/shenyu-spring-boot-starter-plugin/pom.xml
@@ -57,5 +57,6 @@
         <module>shenyu-spring-boot-starter-plugin-uri</module>
         <module>shenyu-spring-boot-starter-plugin-general-context</module>
         <module>shenyu-spring-boot-starter-plugin-mqtt</module>
+        <module>shenyu-spring-boot-starter-plugin-logging-rocketmq</module>
     </modules>
 </project>
